configure.ac: fix behavior of gdrcopy library checking and usage
Currently, gdrcopy library is always checked and used in if it exists,
even if user specified --without-gdrcopy or --with-gdrcopy=no.

This patch adjust beavior of gdrcopy library checking to the following

1. If cuda library is not available, do not check gdrcopy. This is
   because gdrcopy without cuda does not work. If user asked for gdrcopy
   by --with-gdrcopy, fail the configure and print error message.

2. If cuda library is available:
   2a. If user specify --without-gdrcopy or --with-gdrcopy=no, do not
       checkout gdrcopy.
   2b. Otherwise check gdrcopy. If gdrcopy does not exist and user asked
       for gdrcopy, fail and print error message.

Prior to this change, the macro HAVE_GDRCOPY is defined only when
gdrcopy is available.

With this change, the macro HAVE_GDRCOPY is always defined. The value
is used to indicate whether gdrcopy library is available.
Therefore, a change to src/hmem_cuda_gdrcopy.c is made accordingly.

Signed-off-by: Wei Zhang <wzam@amazon.com>