Add the ability to link with clang linker
Currently the clang linker (llvm-link) is not detected as a gnu linker
Libtool doesn't add the appropriate flags to pick up the mapfiles
This detects clang as a linker and will make sure the appropriate link
line is enabled.  Newer versions of autotools may obsolete this detection
which is probably a bit hacky.